requirements.txt
python 3.6
opencv2
numpy